Provide default prices config from json
