+++ price/pricing/config.json
@@ -1,0 +1,18 @@
+{
+  "base_prices": {
+    "residential": {
+      "price_per_hour_usd": 0.00036,
+      "price_per_gib_usd": 0.06
+    },
+    "other": {
+      "price_per_hour_usd": 0.00036,
+      "price_per_gib_usd": 0.06
+    }
+  },
+  "country_modifiers": {
+    "US": {
+      "Residential": 1.5,
+      "Other": 1.2
+    }
+  }
+}
